Pure cosmetics
Split Score.scala into Score + Event
