+++ src/test/resources/com/phasmidsoftware/bridge/director/Newton/Newton Scoring 20241217.csv
@@ -1,0 +1,97 @@
+Newton December 17th 2024,,,
+A DW,,,
+1,N,Margaret Meehan & Sheri Iskenderian,
+2,N,Leslie Greenberg & Annie Getyergun,
+3,N,Carol Rubin & Sara Stone,
+4,N,Lisa & Rick Martin,
+5,N,Marsha Greenstein & Judy Taub,
+6,N,Josh Gahm & Marya Van'T Hul,
+1,E,Susan Stearns & Barbara Linder,
+2,E,Carolyn Drapkin & Art Fine,
+3,E,Marcia Ozarowski & Barbara Bix,
+4,E,Susan Korsmeyer & Missy Mahoney,
+5,E,Sherrill Kobrick & Gerri Taylor,
+6,E,Judy Gelfand & Rho Cherry,
+P,1,1,
+1,-50,,
+2,-120,,
+3,140,,
+
+P,2,2,
+1,-800,,
+2,50,,
+3,590,,
+
+P,3,3,
+1,400,,
+2,150,,
+3,430,,
+
+P,4,4,
+1,110,,
+2,-110,,
+3,-50,,
+
+P,5,5,
+1,150,,
+2,-120,,
+3,420,,
+
+P,6,6,
+1,460,,
+2,-140,,
+3,100,,
+
+P,1,6,
+4,-130,,
+5,140,,
+6,-140,,
+
+P,2,1,
+4,-110,,
+5,630,,
+6,-130,,
+
+P,3,2,
+4,-120,,
+5,620,,
+6,-150,,
+
+P,4,3,
+4,100,,
+5,-100,,
+6,100,,
+
+P,5,4,
+4,100,,
+5,620,,
+6,-110,,
+
+P,6,5,
+4,-180,,
+5,120,,
+6,100,,
+
+P,1,5,
+7,200,,
+8,110,,
+,,,
+P,2,6,
+7,150,,
+8,420,,
+,,,
+P,3,1,
+7,-200,,
+8,-50,,
+,,,
+P,4,2,
+7,-200,,
+8,170,,
+,,,
+P,5,3,
+7,-100,,
+8,420,,
+,,,
+P,6,4,
+7,130,,
+8,170,,
